requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Bcrypt
python-dotenv
bootstrap-flask
email_validator
Pillow